Use uint64_t for serial::check_equal_value
On 32-bit platforms, such as Win32, std::size_t is too small, causing
a narrowing error in the 64-bit self test:

```
    static_assert(check_bswap<uint64_t, 0x201000000000000ull, 0x102>);
```
->
```
In file included from common/include/physfs-serial.h:11,
                 from similar/main/ai.cpp:44:                                                                                                                    common/include/serial.h:528:15: error: static assertion failed
  528 | static_assert(check_bswap<uint64_t, 0x201000000000000ull, 0x102>);                                                                                             |               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~                                                                                         common/include/serial.h:528:15: note: constraints not satisfied
common/include/serial.h:26:9:   required for the satisfaction of 'check_equal_value<((unsigned int)Expected), ((unsigned int)serial::bswap(input))>' [with T = long long unsigned int; Expected = 144396663052566528; input = 258]
common/include/serial.h:528:15: error: narrowing conversion of '144396663052566528' from 'long long unsigned int' to 'unsigned int' [-Wnarrowing]
```

+++ common/include/serial.h
@@ -22,7 +22,7 @@
 
 namespace serial {
 
-template <std::size_t Expected, std::size_t Actual>
+template <std::uint64_t Expected, std::uint64_t Actual>
 concept check_equal_value = (Expected == Actual);
 
 template <typename... Args>
